+++ crates/rolldown_common/src/types/import_kind.rs
@@ -1,0 +1,37 @@
+use std::fmt::Display;
+
+#[derive(Debug, Clone, Copy)]
+pub enum ImportKind {
+  Import,
+  DynamicImport,
+  Require,
+}
+
+impl ImportKind {
+  pub fn is_static(&self) -> bool {
+    matches!(self, Self::Import | Self::Require)
+  }
+}
+
+impl TryFrom<&str> for ImportKind {
+  type Error = String;
+
+  fn try_from(value: &str) -> Result<Self, Self::Error> {
+    match value {
+      "import" => Ok(Self::Import),
+      "dynamic-import" => Ok(Self::DynamicImport),
+      "require-call" => Ok(Self::Require),
+      _ => Err(format!("Invalid import kind: {value:?}")),
+    }
+  }
+}
+
+impl Display for ImportKind {
+  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+    match self {
+      Self::Import => write!(f, "import-statement"),
+      Self::DynamicImport => write!(f, "dynamic-import"),
+      Self::Require => write!(f, "require-call"),
+    }
+  }
+}
