+++ crates/rolldown_common/src/types/import_record.rs
@@ -1,47 +1,9 @@
-use std::fmt::Display;
-
 use rolldown_rstr::Rstr;
 
-use crate::{ModuleIdx, SymbolRef};
+use crate::{ImportKind, ModuleIdx, SymbolRef};
 
 oxc::index::define_index_type! {
   pub struct ImportRecordIdx = u32;
-}
-
-#[derive(Debug, Clone, Copy)]
-pub enum ImportKind {
-  Import,
-  DynamicImport,
-  Require,
-}
-
-impl ImportKind {
-  pub fn is_static(&self) -> bool {
-    matches!(self, Self::Import | Self::Require)
-  }
-}
-
-impl TryFrom<&str> for ImportKind {
-  type Error = String;
-
-  fn try_from(value: &str) -> Result<Self, Self::Error> {
-    match value {
-      "import" => Ok(Self::Import),
-      "dynamic-import" => Ok(Self::DynamicImport),
-      "require-call" => Ok(Self::Require),
-      _ => Err(format!("Invalid import kind: {value:?}")),
-    }
-  }
-}
-
-impl Display for ImportKind {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    match self {
-      Self::Import => write!(f, "import-statement"),
-      Self::DynamicImport => write!(f, "dynamic-import"),
-      Self::Require => write!(f, "require-call"),
-    }
-  }
 }
 
 /// See [ImportRecord] for more details.
