+++ crates/rolldown_common/src/types/mod.rs
@@ -10,6 +10,7 @@ pub mod ecma_ast_idx;
 pub mod entry_point;
 pub mod exports_kind;
 pub mod external_module_idx;
+pub mod import_kind;
 pub mod import_record;
 pub mod importer_record;
 pub mod instantiated_chunk;
